Don't export FS methods by default (#5947)
This continues the removal of automatic exporting of runtime methods, focusing on the filesystem support methods, in particular the ones that are used by loading file packages created separately (that is, created by running the file packager manually, then loading them at runtime, in which case the compiler does not see the filesystem usage, and we must export the things it needs so the package finds them).

This is similar to the previous work in that in -O0 or ASSERTIONS any used method not exported will show an error explaining how to fix things. In addition, for these methods we also suggest using FORCE_FILESYSTEM which will export all the filesystem methods in a convenient way.