Clean up JS optimizer traversal (#5960)
Don't traverse the internal structure of var and object, which led us to horrible workarounds in JSDCE. Skip those internal nodes, so that traversal only traverses actual nodes, as intended, and remove those workarounds.

This was always incorrect, but never noticed until JSDCE since our opts run on asm.js, which doesn't have objects etc. And with JSDCE we temporarily worked around it.

This fix will make future optimizations similar to JSDCE easier, which we need to get the full benefit out of metadce (since it needs to scan the JS).

Start that with removing unused objects in JSDCE, which can now clean up a lot more of the JS runtime when it is unused.