Remove final default runtime exports (#5984)
The last piece of this process, this removes the final default runtime exports, the numeric constants for the allocation modes. We use a getter on Module to provide a useful error in -O0 and ASSERTIONS mode; otherwise this is similar to all the other changes, and like them adds testing for the error message reporting.