metadce: JS function support (#6032)
This adds JS function scanning support to metadce. That means that we can DCE the combined JS+wasm in a way that doesn't treat the entire JS world as a single DCE entity (which we did before), instead we see individual functions, and can collect subgraphs between them and wasm.

An example of something this can clean up are the dynCall/invoke functions. wasm calls invoke in JS, which does a try-catch, which calls back into wasm using dynCall. With this PR, if an invoke is not used, we can see that the dynCall is not used either, and both are removable.

Even better, this can remove runtime methods like abort() that are imported into wasm, and wasm runtime exports like stackAlloc() that are exported from wasm. With this PR, if you compile something minimal like

#include <emscripten.h>

EMSCRIPTEN_KEEPALIVE
int add(int x, int y) {
  return x;
}

then you get something pretty minimal - a 132 byte wasm file! (which still contains a little unnecessary stuff, see #5966)

On more C/C++ hello-world type examples, where the runtime is less noticeable anyhow, this typically shrinks by a few %, e.g. 4% on libc++ hello world's JS.

This PR is mostly straightforward, the one tricky area are dynCalls, since we want to support dynCall('v', ptr), i.e., giving the signature at runtime. This PR will analyze those, if it sees a static string being passed to dynCall, it will consider it a static dynCall, but if it sees something it cannot prove is static, it will add links to all the dynCall_* methods. This happens in practice if you use embind or addFunction, for example.

This PR also adds a test for the wasm backend, if the wasm backend is not the default anyhow, and if fastcomp was built with it. That's the case for me locally, but not the case on the emsdk, so this is skipped on the bots. But it helps me locally, and perhaps we should start to build the wasm backend with fastcomp (maybe after the next llvm merge, though).