Allocate main()'s arguments on the stack (#6058)
The stack is much more efficient and avoids pulling in malloc just for this.

This is a dramatic win for simple hello world type things, where it turns out malloc was in fact pulled in just for this. Hello world using emscripten_log goes from 6,654 bytes of wasm to just 206 (the JS is still the bulk of the combined size, though, anyhow). Hello world using printf also shrinks by around 6K.