fix envvar-related emcc link (#6045)
The `emcc` reference page directs the reader to the source of `emcc` to understand how several environment variables affect `emcc`.  `emcc`, however, is merely a wrapper script for loading `emcc.py`.  While the reader is probably able to understand what the intention is, it would be better to point the reader to the correct place initially.