fix setting of `timezone` by `tzset()`
`timezone` and the value returned by `getTimezoneOffset()` are specified
in the same way, so there's no need to negate the value returned
here.  (Compare the glibc extension field `tm_gmtoff` to `struct tm`,
which is specified as seconds east of UTC, which would require a
negation.)

Fix noticed by @brion, partially addresses #5652.