Implicit archive group fix (#6129)
* factor out a function for scanning groups

We're going to need a copy of this function after processing all files
if there was a --start-group without a matching --end-group.

* permit --start-group without a matching --end-group

GNU ld permits --start-group without a matching --end-group, implicitly
adding the --end-group at the end of the command line.  We should
implement the same behavior.

Fixes #5982.

* add tests for --start-group without --end-group

The previous changes turned this case into a non-error, so the tests
need to change to take this into account.