stop using arg%d in code snippets in render_function
We have the strings for each JS argument name already built, we should
index off those whenever possible, so that we have a single point of
truth for the argument names.  We also switch to using .format() on
strings to make the code clearer when the argument name needs to be
repeated multiple times.