WebAssembly function cast emulation improvements (#6334)
This reworks how function cast emulation works in wasm: it uses the new binaryen pass for that (WebAssembly/binaryen#1468), which converts all indirect calls to use a simple ABI with a fixed number of arguments and i64s for all types. This lets stuff like Python work, which casts away extra arguments, making them technically invalid as wasm indirect calls without emulation.

This fixes #6309 (dynamic linking + wasm + function cast emulation in asm2wasm, which did not work) and also gets function cast emulation working in the wasm backend, enabling the test there. This should also get Python working in the wasm backend (however the test there can't be enabled as it's not compiling from source - I think we just bundled the bitcode because python was just huge).

To get asm2wasm dynamic linking working with function cast emulation, the way we get function pointers is changed: we used to get the exported function and use addFunction to add it to the table dynamically. This can't work with function cast emulation since the export is callable from JS, while function pointers need to use the i64 ABI, so we'd need to convert those at runtime, which is not practical. Instead, asm2wasm provides metadata from the backend about the function pointer for each export, and uses those directly. This is better anyhow because it's simpler and likely closer to what proper wasm dynamic linking will use eventually; also, using addFunction is tricky here since we don't know the type, so that couldn't have worked in the wasm backend anyhow.