fix error message about ill-formed js library (#6271)
We load js-libraries one after the other, and had a bug where we assigned to `processed` in the last iteration, failed to eval the current one, and the error logging printed out the old value, so the error was reported about the previous library, and not the one with the invalid code in it. The fix is to assigned `undefined` to that var at the top of each iteration.