Updated the fetchRemotePackage function
If the Browse is Chrome/Safari/IE the file needs to be served using SimpleHTTPServer or the like. If the html file is run in these browsers, the XHR script throws an error. The XHR open call has been replaced with am if-else block that checks for the browser and protocol and gracefully exits.