when assertions are on, show a warning when we try to grow memory above the practical limit (#6568)
The practical limit is 2GB and most (all?) browsers don't allow allocating more now, so emscripten doesn't ask for more, and instead increments memory in small chunks towards that limit (hoping it will be enough). This adds a warning when we reach the limit there, to make this less surprising to users.