Simpliy use of Configuration object emscripten.py
I'm not sure any emscripten.py has its own Configuration.  The
regular one at `shared.configuration` is already created on import
and is identical.  So this change basically reduces the number
of Configuration (and also TempFiles which configuration creates)
objects in the system.

Also rename `keywords` to `options` which is more common for
the result of opt_parse.