Add 'main' symbol to test_fs_dict
Without this lld will fail since this test includes main
in EXPORTED_FUNCTIONS.  It seems that the existing linkers allow
for symbols in EXPORTED_FUNCTIONS to be missing (?).

Since this is the only test that seems to rely on this behavior
I'm tempted to say the failing at link in this case is actually the
correct behavior, and fix the test to include a `main`.