warn on native object files in .a (#6618)
A potentially common pitfall is to have native object files there, if it was build normally and not using emcc. Before this we didn't warn, since we saw llvm-nm passed and we only checked that, and then the file was silently skipped, which could be confusing.

Note that sometimes .a files contain extra things, like Rust places metadata files in them. We need to just skip over those, basically, so we can link in the relevant code. So we can warn here, but intentionally do not throw an error.