Handle debug flags in lld path
Pass --strip-debug to lld when not building in debug mode.
Also pass -g to wasm-emscripten-finalize in debug mode.

This fixes the test_bswap64 test which was crash in wasm-opt
because dynCall symbols were being created both by
wasm-emscripten-finalize and then again in wasm-opt.

There is an underlying but with wasm-opt and fpcast-emu generating
duplicate functions which can be addressed separately.