Set visibility with EMSCRIPTEN_KEEPALIVE so that lld exports the symbol. (#6625)
In the previous s2wasm path GC was don't only by llvm-link and it
honors the used attribute.  Then s2wasm I think will export any
remaining functions.

With the lld path we have GC going on llvm-link (for now) and then
again in lld.  Function are exported if they are explicitly exported
from the command line or if they have default visibility.

Alternatively we could have -fvisibility=default when we compile?