Move `environ` to C code, to work around lld limitations (#6614)
Similar to the recent tzname etc. changes. However, environ is trickier because we want the environment to be read and written from JS and C, and in particular initialized from C (since C knows when it exists). So we need a constructor to initialize the C environ object, by calling into JS.

The constructor can't be dce'd out, so I moved the extras into a separate system library, libc-extras, so it isn't linked in unless necessary.

The problem with this PR is that we depend a lot more on deps_info.json, that is, not on normal linking. We need to know if a function will include something that includes, through C or JS, use of the environment.

Intended followups:

 * The more we use deps_info.json, the more we depend on something fairly fragile. We should at least improve warning and error handling there.
 * We should probably adjust the constructor priority of that new libc-extras constructor, need to look into best practices there.