stress the importance of the 'clamp' trap mode [ci skip] (#6655)
After the recent LLVM upgrade, several reports have come in of people seeing traps where they didn't before, so I guess the optimizer now generates code more prone to this. This PR updates the docs to explain that better, include the error messages for better searchability, and stresses the importance of using clamp.

We should perhaps consider changing the default to clamp. I know we had a long discussion on this last time, though.