Fix flake8 warnings in shared.py (#6786)
One non-trival change here is the way global variables in the
config file are checked for.  Previously we were using a strange
try/catch technique to detect variables in the global scope.
With this change we simply initialize the globals before reading
the config and then check them against None.