Move run_js helper to tests/runner.py
This wrapper was only used by test code so move it there.

Rename it to run_js_default to make it more clear what it does.

This is part of push to separate the js runner code used during
testing from the JS running code used by the actual toolchain.

Prefer run_js_tool for running JS are part of the actual toolchain.