Always run via subprocess in python-selector. (#6795)
If we run via import then we don't execute top level code that
checks for 'if __name__ == "__main__"'.  This means the scripts
behave differently depending on which method the selector chooses.

Also, since we support both 2 and 3 for now we might as well
always run under `sys.executable` which we know for sure exists
(unlike `python2` which doesn't always exist on mac for example).