Show an error with explanation when a user compiles to LLVM bitcode with SIDE_MODULE (#6835)
We should not silently ignore the SIDE_MODULE flag there, as it indicates the user thought the command was emitting the final shared library. Instead, the bitcode needs to be compiled to the final shared library with that flag, by emitting a .wasm or .js file (for a wasm or asm.js shared library, respectively).