enable most GL tests in CircleCI (#6816)
* make browser.test_float_tex compatible with Mesa software backends

For some reason, glGetShaderInfoLog() always returns the following
information log although the compile operation was successful:

  0:2(12): warning: extension `GL_ARB_gpu_shader5' unsupported in vertex shader

and it seems that the mentioned extension is not well supported by Mesa
software backends yet. The test can still work as expected though, as
long as only the compile status is checked.

* split browser.test_html5_webgl_create_context into two

Anti-aliasing is not well supported by the software renderer in Mesa,
and this breaks browser.test_html5_webgl_create_context. It's suboptimal
to completely skip this test, so use a trimmed variant as a workaround.

* use Firefox + headless X server for browser tests in CI

This is the prerequisite to enable tests requiring GL. Firefox does not
support WebGL in headless mode yet.

Travis CI configuration is not updated because of time limit concerns.

* actually enable GL tests in CircleCI

Tests requiring certain hardware access are now classified in finer
granularity, inspired by the initial idea and implementation here:

https://github.com/kripken/emscripten/pull/5901#discussion_r160934596

The migration is quite straightforward, except that
browser.test_html5_webgl_create_context is skipped due to the lack of
proper anti-aliasing support.