Add preload plugin to compile wasm side modules async (#6663)
This is a solution to #6633 related to dlopen'ing SIDE_MODULES using async wasm compilation (required by Chromium).

This adds a new preload handler for .so files containing wasm dynamic libraries, to compile them using the async API (WebAssembly.instantiate) at data-loading time. The actual dynamic linking happens synchronously when calling dlopen later. The downside here is that you may end up compiling wasm that never actually gets dlopen'ed, but I think that's probably a domain-specific problem and will be up to the user to resolve, ultimately.

The only change required on the user end is to add --use-preload-plugins when file packaging.

Of note is that loadWebAssemblyModule most be called one at a time, since it sets up an environment in global variables that is later used when the wasm is compiled asynchronously. So this uses a chain of promises to do the compilation, rather than just Promise.all.