Require all EXPORTED_FUNCTIONS to be implemented
For asm2wasm this means turning the existing warning into an
error.

For wasm backend the means adding the '-u' flag to the linker
which will cause the linker to treat the symbol as undefined
and search for it in library files.   If the symbol is not found
the linker will error out.

This is a small behavior change but I can't think of any users
who would not want this behavior.  Quite the contrary, if you
add something to EXPORTED_FUNCTIONS and its not found, you
probably want to know at compile time.