[Test Runner] Separate the two types of `post_build` option (#6853)
Previously post_build would either be a string or a tuple and
represented two different kind of callback.  Since there use
is distinct (no callers were using both) it is a lot more clear
to separate them.

This also allows more tests to go threw the fast path in the
`build` method.

As a followup we can probably remove the js_transform completely
perhaps.