misc testing/CI tidy-up (#6890)
* remove browser overriding from the interactive test suite

The logic has already been in tests/runner.py.

* mute a few test warnings

* turn on Firefox OffscreenCanvas support in CircleCI

OffscreenCanvas is behind the gfx.offscreencanvas.enabled
preference. The relevant tests are silently reported as ok if the
feature is not enabled. Unfortunately, these tests are still skipped
because of several bugs.

* enable browser.test_sdl_ttf_render_text_solid on CircleCI

ttf-mscorefonts-installer allows for easy installation of the Microsoft
True Type Core Fonts for the Web including Arial. The package requires
acceptance of an EULA*. Informally speaking:

  These fonts were provided by Microsoft "in the interest of
  cross-platform compatibility". This is no longer the case, but they
  are still available from third parties.

  You are free to download these fonts and use them for your own use,
  but you may not redistribute them in modified form, including changes
  to the file name or packaging format.

* http://corefonts.sourceforge.net/eula.htm