asm.js validation testing fix (#6995)
Our check for asm.js validation in spidermonkey is just a string check for a warning with "asm.js" in it. This hits a false positive in the warning on addFunction in the wasm backend (the warning mentions that the type is not needed in asm.js). To avoid this, the check can see if we are even in wasm mode, and if not, if we expect asm.js to validate - if no to either of those, there's nothing to check.

The problem was not noticed on the bots as we don't run spidermonkey there.