Fix and enable other.test_bad_triple (#7021)
This was not tested on the bots, and we didn't notice that by mistake we refactored the condition to the opposite of the expected (it should warn about a bad triple, not return an error code).

After fixing that, I also managed to fix it on the bots: the test builds - natively - a cpp to bitcode, then checks we warn about the triple when we compile such a native bitcode to js/wasm. The problem on CI was that stddef.h is not present, that is, native building doesn't work using clang there for some reason. Not sure why, but in any case, we don't need that, I made the test just build an empty cpp file, which is enough.