Support loading a global or function in a side module that is provided by a later side module (#6963)
In asm.js we handle this by the side module's JS creating indirection functions (so that when the property is actually used later, after loading the other side module, we find it). In wasm, though, side modules are pure wasm and cannot help in their own loading in such a manner.

To get around that, this adds a JS proxy on the env object, which sees when the side module tries to load a missing property. This can handle both global helper functions (g$ prefix in the current side module model), and regular functions.

Fixes  #6951