Add an assertion for checking the .buffer after requesting a WebAssembly.Memory that is shared (#7023)
If a browser has SharedArrayBuffer but not wasm-threads, then the call to create the Memory will simply ignore the "shared" property. Checking if the output is an SAB is therefore a way to check if the browser was wasm-threads support (if not, the user may need to flip a flag).