Deprecate emscripten_has_threading_support
This is a followup to the removal of USE_PTHREADS=2 (#7007).

Threading support is something we set at compile time and
can be detected with __EMSCRIPTEN_PTHREADS__.  If threading
is not support in the browser the resulting module will not
load so that is no need for a runtime detection feature.