Wasm-only mode is not a requirement for async compilation (#7036)
Fixes #5567.

To be conservative, we disabled async compilation any time that asm2wasm was not emitting wasm-only code. That meant that if we had a fallback to asm.js, or used an asm.js feature like the emterpreter, async compilation was disabled. Things like EMULATE_FUNCTION_POINTER_CASTS also disable wasm-only mode for now (but perhaps shouldn't). However, this wasn't an actual hard limitation - turns out that just some tests needed to be updated. (Where we do need to disable async compilation is in things like using the binaryen interpreter to run the wasm, as there we do need to synchronously load the wasm, at least in how the current interpreter integration works - low priority, anyhow.)