Remove redundant test: other.test_no_nuthin_2
This test was trying to verify that setting EXPORTED_RUNTIME_METHODS
to [] would save space.  In fact this is the default now so this
setting has no effect.  The size differences were purely a result of
the filename being embedded in the JS in a few places.

In fact, the 'no_nuthin' output was larger than the `normal` output
because of this, but the test was only testing the absolute size
difference.