Support invokes in wasm dynamic libraries even if the main file is missing that signature (#7049)
We use invoke_[sig] to call out to JS, do a try-catch, and so handle catching exceptions. There is a separate function for each signature. A potential problem is that a dynamic library may have a signature not present in the main file, so in asm.js each library adds invoke functions as needed. In wasm, however, a library is a pure wasm file, so it doesn't have JS that can help with loading itself.

To fix that, this PR properly implements the invoke_X handler (which was always present for future use, but not actually implemented). That does a generic invoke regardless of the number of arguments, using the JS arguments object. That's not very efficient, but it works for now until we get proper exception handling in wasm. This builds on the work in #6963 for using a JS Proxy to detect missing imports to a loaded wasm library (we detect a missing invoke_* import, and hook it up to invoke_X - in theory we could use new Function to create a specialized one for it).

See #6963 (comment)