Don't run tests in CANONICAL_TEMP_DIR
This way the EMTEST_SAVE_DIR directory and the canonical temp
directory used by EMCC_DEBUG=1 are distinct.

This fixes a specific issue with the @uses_canonical_tmp decorator
that tries to clean the canonical_tmp_dir.

We can also simplify runner.clear() since we know that CWD will always
be different to the emscripten temp dir.