Simplify NO_* options (#7151)
We have a few options starting with NO_, such as NO_EXIT_RUNTIME. The double negative when writing something like NO_EXIT_RUNTIME=0 can be confusing. This PR simplifies things by making NO_ an optional prefix that flips a boolean option, that is,

-s EXIT_RUNTIME=1

is the same as

-s NO_EXIT_RUNTIME=0.

As this is optional, no breakage should occur for existing users.