Remove explicit dependency on pthread_mutux functions from library_fetch.js (#7195)
These depedencies are currently causing warning in all builds with
FETCH=1+USE_PTHTREADS=1 as far as I can tell.  This is because when
pthreads are enabled these function are implemented in native code
in musl and are named with a leading `__`.

In fact, I far can tell the fetch library doesn't actually depend
on these functions so simply removing the dependency fixes the
issue.