Add --export-dynamic to lld args
I recently made a change to lld to disables this behaviour
by default: https://reviews.llvm.org/D52587

Sadly this broke a few things for emscripten. For example, it
meant that EMSCRIPTEN_KEEPALIVE was not longer enough to save
a native code symbol from GC.

For now add --export-dynamic to the lld args to restore the
previous behavior.