Cleanup scons support and remove config file parsing
This change is precursor to removing EMSCRIPTEN_ROOT from
the emscripten config file.  Now scons requires EMSCRIPTEN_ROOT
to be set explicitly in the environment.

The rational for this change is that having the EMSCRIPTEN_ROOT
set in the config can leave is in a contradictory state.  e.g.
when emcc runs from location X but the config file says EMSCRIPTEN_ROOT
is location Y.  In most cases it is the tool entry point that dictates
the root directory.

The only exception was scons which was trying to imply the root
based on the config.

Also importantly this change removes the `eval` of the config file from
scons which should allow is to be more strict in how we parse and
evaluate the config file in `tools/shared.py` since there is now only
one consumer.