Remove EMSCRIPTEN_ROOT from config file
This was previously used by external tools to find emscripten.
A better way to do that is to look for `emcc` in PATH or
check for EMSCRIPTEN_ROOT in the environment.

Hopefully once we are the only parser of our config file
we can nail it down a little and remove the `exec`.
As well as being cleaner this will also allow the config file
to know where it is via __file__ which allows for the config
to include relative paths(useful for installing toolchains
with fix local config files).