Fix a crash if the worker has the 'window' object defined (#7231)
Fixes #7222

In some situations, the window object may be defined in a worker context, for example if window is defined for a polyfill. In these cases, the worker would report ENVIRONMENT_IS_WEB true and attempt to access document.currentScript. This would cause the worker to crash.

This PR inverts that check so that we check if we're a worker rather than web. This shouldn't cause any changes to functionality.