Support input from the global Module object in MODULARIZE_INSTANCE (#7293)
fixes #7101

Background: normally, if you define Module and properties on it, like preRun, then we notice that and use them. In MODULARIZE, however, the user creates the instances of the module, and can pass Module or something else as desired (if we passed Module there it might be surprising). However, in MODULARIZE_INSTANCE mode we are more like the normal mode in that there is a single instance, started automatically, and defining things on Module is how the user can influence it.