Cleaner handling of clusure compiler errors
Previously you would see an emscripten internal backtrace
if closure compiler failed.

Sadly we have a lot of existing warnings so we don't display
the process stderr unless we the subprcess returns non-zero.

I guess we should try to address the closure warnings, at least
the ones in emscripten proper.