[wasm-backend] Move fixImports from runtime to compile time
With the wasm backend with don't have leading "_" on symbol
names but to fit in with the current codebase we pretend that
we do right up until the end.

This change moves the stripping on the '_' from runtime, right
before instantiate to compile time.  This also allows us to
add error checking to avoid duplicate dictionary keys that
could result.