Waterfall CI, test upstream LLVM wasm backend (#7474)
This gets CI for the upstream LLVM wasm backend on Circle. It uses the emsdk to get the lkgr build from the waterfall (using the code @jgravelle-google added), then uses that for the LLVM, clang, and Binaryen binaries.

This uses Python3 and found a bunch of bugs there, fixes already in separate PRs that have landed. This PR includes the same tests currently run on the waterfall (binaryen2, wasmobj2, other) and they all pass.