if many browser tests hang, skip running the browser portion (#7486)
In the browser test suite we wait for tests to finish or a timeout to occur. If the test just hangs (say, waiting on some event that never happens) then we hit the timeout. This doesn't happen normally, but if there's a problem on CI, like we have temporary breakage of some form, then we can hit the timeout on every single test, and we have hundreds of them which is very painful.

To avoid that, this PR tracks hanging/unresponsive tests, and after we reach 10 we stop invoking the browser and show a message instead (we still test compilation though, which is safe).

*    This should not hide errors as it will only skip after 10 of them - enough that we should be paying a lot of attention to the results anyhow!
*    In practice, I've never seen 10 random timeouts occur - maybe 1 or 2, not sure if even 3 or 4. So we should not hit this randomly, and only when things are just broken for that test run anyhow.